Add basic test suite for endpoint_router
