+++ package/endpoint_router/endpoint_router/src/endpoint_router_load.c
@@ -588,3 +588,78 @@ error:
 
   return NULL;
 }
+
+static void filters_destroy(filter_t **filter_loc)
+{
+  if (filter_loc == NULL || *filter_loc == NULL) {
+    return;
+  }
+  filter_t *filter = *filter_loc;
+  filter_t *next = NULL;
+  while (filter != NULL) {
+    next = filter->next;
+    if (filter->data != NULL) free(filter->data);
+    free(filter);
+    filter = next;
+  }
+  *filter_loc = NULL;
+}
+
+static void forwarding_rules_destroy(forwarding_rule_t **forwarding_rule_loc)
+{
+  if (forwarding_rule_loc == NULL || *forwarding_rule_loc == NULL) {
+    return;
+  }
+  forwarding_rule_t *forwarding_rule = *forwarding_rule_loc;
+  forwarding_rule_t *next = NULL;
+  while (forwarding_rule != NULL) {
+    next = forwarding_rule->next;
+    if (forwarding_rule->dst_port_name != NULL && forwarding_rule->dst_port_name[0] != '\0') {
+      free((void *)forwarding_rule->dst_port_name);
+    }
+    filters_destroy(&forwarding_rule->filters_list);
+    free(forwarding_rule);
+    forwarding_rule = next;
+  }
+  *forwarding_rule_loc = NULL;
+}
+
+static void ports_destroy(port_t **port_loc)
+{
+  if (port_loc == NULL || *port_loc == NULL) {
+    return;
+  }
+  port_t *port = *port_loc;
+  port_t *next = NULL;
+  while (port != NULL) {
+    next = port->next;
+    if (port->name != NULL && port->name[0] != '\0') {
+      free((void *)port->name);
+    }
+    if (port->pub_addr != NULL && port->pub_addr[0] != '\0') {
+      free((void *)port->pub_addr);
+    }
+    if (port->sub_addr != NULL && port->sub_addr[0] != '\0') {
+      free((void *)port->sub_addr);
+    }
+    pk_endpoint_destroy(&port->pub_ept);
+    pk_endpoint_destroy(&port->sub_ept);
+    forwarding_rules_destroy(&port->forwarding_rules_list);
+    free(port);
+    port = next;
+  }
+  free(port);
+  *port_loc = NULL;
+}
+
+void router_teardown(router_t **router_loc)
+{
+  if (router_loc == NULL || *router_loc == NULL) {
+    return;
+  }
+  router_t *router = *router_loc;
+  if (router->name != NULL) free((void *)router->name);
+  ports_destroy(&router->ports_list);
+  free(router);
+  *router_loc = NULL;
+}
